[FEATURE] Switched to 100% composer autoloader.
Note: it's a bit slower (from 20ms to > 35ms) unless you add XCache or OPCache, then it's actually less slow. It's all depending on how many dependencies you have installed though.
