+++ composer.json
@@ -25,6 +25,7 @@
         "psr-4": {
             "Pecee\\": "src/"
         }
+        "files": ["bootstrap.php"]
     },
     "config": {
         "preferred-install": "dist"
